Add get fetch for stage
